WIP: Add UI shell
