--- src/devtools/Panel-new.svelte
@@ -1,14 +1,0 @@
-<script lang="ts">
-  import './panel/panel.css'
-</script>
-
-<div
-  class="min-h-screen min-w-screen grid grid-rows-[30px_auto] text-white bg-black"
->
-  <div class="p-2 border border-[#333437]">Webnative</div>
-  <div class="grid grid-cols-[1fr_1fr_2fr] divide-x divide-[#333437]">
-    <div class="p-2">app namespace column</div>
-    <div class="p-2">event list here</div>
-    <div class="p-2">payload here</div>
-  </div>
-</div>
