+++ src/devtools/panel/Namespaces.svelte
@@ -1,0 +1,8 @@
+<div class="grid grid-rows-[19px_auto]">
+  <div
+    class="flex flex-col justify-center px-4 py-1 bg-[#262626] text-[9px] uppercase"
+  >
+    Active namespaces
+  </div>
+  <div class="px-4 py-2">Namespace entries</div>
+</div>
